% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{C2_1day}
\alias{C2_1day}
\title{Calculates the mean, standard deviation and C2 metric value taking a 1 day lag}
\usage{
C2_1day(data, t)
}
\arguments{
\item{data}{array of adjusted number of daily confirmed cases}

\item{t}{Day number}
}
\value{
Returns the mean, standard deviation and C2 metric value
}
\description{
This is similar to the function C2, but it uses the data of seven days and a 1-day lag.
}
