% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{C_metric}
\alias{C_metric}
\title{Returns the value of C1, C2 or C3 metric and the corresponding limit}
\usage{
C_metric(bound_metric, df, t)
}
\arguments{
\item{bound_metric}{C1, C2, or C3 metric can be selected}

\item{df}{Dataframe with two columns which contains dates and corresponding number of daily confirmed cases}

\item{t}{Day number}
}
\value{
Returns the mean, standard deviation, threshold value for that metric and the metric value
}
\description{
Calls C1, C2, C3 functions and returns the respective metric value and the threshold limit.
}
\note{
This function is called in the "threshold" function
}
