% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{Cmetric}
\alias{Cmetric}
\title{Returns the value of C1, C2 or C3 metric and the corresponding limit}
\usage{
Cmetric(bound_metric, num_adj, i)
}
\arguments{
\item{bound_metric}{C1, C2, or C3 metric can be selected}

\item{num_adj}{array of adjusted number of daily confirmed cases}

\item{i}{Day number}
}
\value{
Returns the mean, standard deviation, threshold value for that metric and the metric value
}
\description{
Calls C1, C2, C3 functions and returns the respective metric value and the threshold limit.
}
