% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{compare_results}
\alias{compare_results}
\title{Identifies method with least error}
\usage{
compare_results(
  population = 18710922,
  gamma = 1/14,
  cur_day = 100,
  start_date = "2020-3-13",
  last_n_day = 20,
  last_limit = 30,
  next_n_days = 20,
  data,
  ub_for_adjustment = 5,
  bound_metric = "C3_1day",
  df_confirmed_values,
  mu
)
}
\arguments{
\item{population}{number of people in the state}

\item{gamma}{recovery rate}

\item{cur_day}{current day number for start of prediction phase}

\item{start_date}{start date in the considered dataset}

\item{last_n_day}{number of days in training phase}

\item{last_limit}{maximum number of days in the validation period}

\item{next_n_days}{number of days in the prediction phase}

\item{data}{state wise daily cases adjusted data for the given state}

\item{ub_for_adjustment}{upper bound for the duration of a jump or drop}

\item{bound_metric}{C1, C2, or C3 metric can be selected}

\item{df_confirmed_values}{Dataframe of dates and observed number of daily confirmed cases}

\item{mu}{mortality rate of the infection}
}
\value{
Returns graph showing observed and predicted cummulative cases using the best method and the corresponding MSE for validation and prediction period.
}
\description{
Compares the validation rmse using original and adjusted data and returns the one with lesser error.
}
\note{
This function is called by the user. It prints the MSE values of predictions made with and without making the adjustments and returns the best predictions.
}
