% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{comparing_results}
\alias{comparing_results}
\title{Compares the validation rmse using original and adjusted data and returns the one with lesser error}
\usage{
comparing_results(output_original, output_adjusted)
}
\arguments{
\item{date}{array of dates}

\item{num}{array of observed number of daily confirmed cases}

\item{num_adj}{array of adjusted number of daily confirmed cases}
}
\value{
graph showing observed and predicted cummulative cases using the best method
}
\description{
Compares the validation rmse using original and adjusted data and returns the one with lesser error
}
