% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{plot_adjustment}
\alias{plot_adjustment}
\title{plots graph of daily cases with and without adjustment}
\usage{
plot_adjustment(date, num, num_adj)
}
\arguments{
\item{date}{array of dates}

\item{num}{array of observed number of daily confirmed cases}

\item{num_adj}{array of adjusted number of daily confirmed cases}
}
\value{
graph showing number of confirmed daily cases with and without adjustment
}
\description{
plots graph of daily cases with and without adjustment
}
