% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{plot_cumulative}
\alias{plot_cumulative}
\title{Plots cumulative cases graph for original and adjusted predictions}
\usage{
plot_cumulative(output_original, output_adjusted)
}
\arguments{
\item{output_original}{output of sisd_cummulative for original state-wise data}

\item{output_adjusted}{output of sisd_cummulative for adjusted state-wise data}
}
\value{
graph showing observed, trained and predicted values for both original and adjusted data
}
\description{
Provides a visual comparison between predictions made with and without the adjustments and the observed number of cases.
}
