\name{sisd_cummulative}
\alias{sisd_cummulative}
\title{Prediction of Cumulative number of cases using data driven modified SIS model}
\description{Prediction of Cumulative number of cases using data driven modified SIS model}
\usage{sisd_cummulative(
   population,
   gamma,
   cur_day,
   last_n_day,
   last_limit,
   next_n_days,
   data,
   adjusted,
   obs_arr,
   num_adj,
   mu)
  }
\value{graph showing the training phase and prediction of cumulative number of cases}
