\name{threshold}
\alias{threshold}
\title{Identifies jumps and drops using given metric and adjusts if the length of the jump or drop is less than the upper bound}
\description{Identifies jumps and drops using given metric and adjusts if the length of the jump or drop is less than the upper bound}
\usage{threshold(
           ub_for_adjustment,
           bound_metric,
           num,
           num_adj)
  }
\value{adjusted data}
