% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{C1}
\alias{C1}
\title{Calculates the mean, standard deviation and C1 metric value}
\usage{
C1(df, t)
}
\arguments{
\item{df}{Dataframe with two columns which contains dates and corresponding number of daily confirmed cases}

\item{t}{Day number}
}
\value{
Returns the mean, standard deviation and C1 metric value
}
\description{
The C1 uses moving sample average and sample standard deviation
to standardize each observation. It calculates the sample average and
sample standard deviation using the data of 7 days prior to the current day.
}
\note{
This function is called by the "C_metric" function which in turn is called in the "threshold" function
}
