% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{C1}
\alias{C1}
\title{Calculates the mean, standard deviation and C1 metric value}
\usage{
C1(data, t)
}
\arguments{
\item{data}{array of adjusted number of daily confirmed cases}

\item{t}{Day number}
}
\value{
Returns the mean, standard deviation and C1 metric value
}
\description{
The C1 uses moving sample average and sample standard deviation
to standardize each observation. It calculates the sample average and
sample standard deviation using the data of 7 days prior to the current day.
}
