% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{C2}
\alias{C2}
\title{Calculates the mean, standard deviation and C2 metric value taking a 2 day lag}
\usage{
C2(df, t)
}
\arguments{
\item{df}{Dataframe with two columns which contains dates and corresponding number of daily confirmed cases}

\item{t}{Day number}
}
\value{
Returns the mean, standard deviation and C2 metric value
}
\description{
The C2 uses moving sample average and sample standard deviation
to standardize each observation. It calculates the sample average and
sample standard deviation using the data of seven days and a 2-day lag.
}
\note{
This function is called by the "C_metric" function which in turn is called in the "threshold" function
}
