% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{C3}
\alias{C3}
\title{Calculates the mean, standard deviation and C3 metric value taking a 2 day lag}
\usage{
C3(df, t)
}
\arguments{
\item{df}{Dataframe with two columns which contains dates and correspoding number of daily confirmed cases}

\item{t}{Day number}
}
\value{
Returns the mean, standard deviation and C3 metric value
}
\description{
The C3 uses information from the C2 statistic of the past two days.
}
\note{
This function is called by the "C_metric" function which in turn is called in the "threshold" function
}
