\name{comparing_results}
\alias{comparing_results}
\title{Identifies jumps and drops using given metric and adjusts if the length of the jump or drop is less than the upper bound}
\description{Identifies jumps and drops using given metric and adjusts if the length of the jump or drop is less than the upper bound}
\usage{comparing_results(
           output_original,
           output_adjusted)
  }
\value{output}
