% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{comparing_results}
\alias{comparing_results}
\title{Identifies method with least error}
\usage{
comparing_results(output_original, output_adjusted)
}
\arguments{
\item{output_original}{output of sisd_cummulative for original state-wise data}

\item{output_adjusted}{output of sisd_cummulative for adjusted state-wise data}
}
\value{
graph showing observed and predicted cummulative cases using the best method
}
\description{
Compares the validation rmse using original and adjusted data and returns the one with lesser error.
}
