% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{plot_adjustment}
\alias{plot_adjustment}
\title{plots graph of daily cases with and without adjustment}
\usage{
plot_adjustment(df_confirmed_values, df_adjusted_confirmed_values)
}
\arguments{
\item{df_confirmed_values}{dataframe with two columns which contains dates and corresponding observed number of daily confirmed cases}

\item{df_adjusted_confirmed_values}{dataframe with two columns which contains dates and corresponding adjusted number of daily confirmed cases}
}
\value{
Returns the graph showing number of confirmed daily cases with and without adjustment
}
\description{
Visual representation of the adjustments made using a given metric.
}
\note{
This function is called in "sisd_cummulative"
}
