% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{sisd_cummulative}
\alias{sisd_cummulative}
\title{Predict with or without adjusting the data}
\usage{
sisd_cummulative(
  population = 18710922,
  gamma = 1/14,
  cur_date = "2020-5-29",
  start_date = "2020-3-13",
  last_n_day = 20,
  last_limit = 30,
  next_n_days = 20,
  data,
  adjusted = 0L,
  ub_for_adjustment = 5,
  bound_metric = "C3_1day",
  df_confirmed_values,
  mu
)
}
\arguments{
\item{population}{number of people in the state}

\item{gamma}{recovery rate}

\item{cur_date}{current date for start of prediction phase}

\item{start_date}{start date in the considered dataset}

\item{last_n_day}{number of days in training phase}

\item{last_limit}{maximum number of days in the validation period}

\item{next_n_days}{number of days in the prediction phase}

\item{data}{state wise daily cases adjusted data for the given state}

\item{adjusted}{predictions can be made with and without adjustment in data}

\item{ub_for_adjustment}{upper bound for the duration of a jump or drop}

\item{bound_metric}{C1, C2, or C3 metric can be selected}

\item{df_confirmed_values}{Dataframe of dates and observed number of daily confirmed cases}

\item{mu}{mortality rate of the infection}
}
\value{
Returns graph showing the training phase and prediction of cummulative number of cases, data frame of cummulative cases in training and prediction phase, mean square error of validation and prediction period
}
\description{
Prediction of Cumulative number of cases using data driven modified SIS model
}
\note{
This function is called in the function "compare_results" It calls the function "plot_adjustment" for a visual depiction of the adjustments made by the given metric.
}
