% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{sisd_cummulative}
\alias{sisd_cummulative}
\title{Prediction of Cumulative number of cases using data driven modified SIS model}
\usage{
sisd_cummulative(
  population,
  gamma,
  cur_day,
  last_n_day,
  last_limit,
  next_n_days,
  data,
  adjusted,
  obs_arr,
  num_adj,
  mu
)
}
\arguments{
\item{population}{number of people in the state}

\item{gamma}{recovery rate}

\item{cur_day}{current day number for start of prediction phase}

\item{last_n_day}{number of days in training phase}

\item{last_limit}{maximum number of days in the validation period}

\item{next_n_days}{number of days in the prediction phase}

\item{data}{state wise daily cases adjusted data for the given state}

\item{adjusted}{predictions can be made with and without adjustment in data}

\item{obs_arr}{array of observed number of daily confirmed cases}

\item{num_adj}{array of adjusted number of daily confirmed cases}

\item{mu}{mortality rate of the infection}
}
\value{
graph showing the training phase and prediction of cumulative number of cases, data frame of cumulative cases in training and prediction phase,
array of observed and predicted cumulative cases, mean square error of validation and prediction period
}
\description{
Prediction of Cumulative number of cases using data driven modified SIS model
}
