% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPredSIS.R
\name{threshold}
\alias{threshold}
\title{Identifies and adjusts jumps and drops}
\usage{
threshold(ub_for_adjustment, bound_metric, df_confirmed_values)
}
\arguments{
\item{ub_for_adjustment}{upper bound for the duration of a jump or drop}

\item{bound_metric}{C1, C2, or C3 metric can be selected}

\item{df_confirmed_values}{Dataframe of dates and observed number of daily confirmed cases}
}
\value{
Returns dataframe of dates and adjusted data using given metric
}
\description{
Identifies jumps and drops using given metric and adjusts if the length of the jump or drop is less than the upper bound.
}
\note{
This function is called in "sisd_cummulative"
}
