% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisd.R
\name{threshold}
\alias{threshold}
\title{Identifies jumps and drops using given metric and adjusts if the length of the jump or drop is less than the upper bound}
\usage{
threshold(ub_for_adjustment, bound_metric, num, num_adj)
}
\arguments{
\item{ub_for_adjustment}{upper bound for the duration of a jump or drop}

\item{bound_metric}{C1, C2, or C3 metric can be selected}

\item{num}{array of observed number of daily confirmed cases}

\item{num_adj}{Copy of num for adjustment}
}
\value{
adjusted data using given metric
}
\description{
Identifies jumps and drops using given metric and adjusts if the length of the jump or drop is less than the upper bound
}
